+++ src/main/java/group3/kino/movieAdministration/controller/MovieAdminController.java
@@ -1,11 +1,10 @@
-package group3.kino.controller;
+package group3.kino.movieAdministration.controller;
 
-import group3.kino.model.Movie;
-import group3.kino.service.MovieAdminService;
+import group3.kino.movieAdministration.model.Movie;
+import group3.kino.movieAdministration.service.MovieAdminService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.Optional;
@@ -21,7 +20,7 @@ public class MovieAdminController {
         this.movieAdminService = movieAdminService;
     }
 
-   /* @PostMapping("/addMovie")
+   @PostMapping("/addMovie")
     public ResponseEntity<Movie> addMovie(@RequestBody Movie movie) {
         movieAdminService.save(movie);
         return new ResponseEntity<>(movie, HttpStatus.OK);
@@ -52,8 +51,5 @@ public class MovieAdminController {
             return new ResponseEntity<>(movie_, HttpStatus.OK);
         }
         return new ResponseEntity<>(null, HttpStatus.BAD_REQUEST);
-
     }
-
-    */
 }
