+++ src/main/java/group3/kino/movieAdministration/service/IMovieAdminService.java
@@ -1,0 +1,7 @@
+package group3.kino.movieAdministration.service;
+
+import group3.kino.movieAdministration.model.Movie;
+
+
+public interface IMovieAdminService extends CRUD<Movie, Long> {
+}
