--- src/main/java/group3/kino/service/IMovieAdminService.java
@@ -1,8 +1,0 @@
-package group3.kino.service;
-
-import group3.kino.model.Movie;
-import org.springframework.stereotype.Service;
-
-
-public interface IMovieAdminService extends CRUD<Movie, Long> {
-}
